TTL no longer needs to be constant.
When using a constant TTL, some applications will "hang" every TTL seconds due to a massive invalidation/reprocessing.

Instead, it is often best to spread the invalidations during a large time window, so that cache invalidations are distributed evenly during the application execution;
E.g., Instead of keeping everything for exactly 2h, it's probably a good idea to keep each cached value between 1h30 and 2h30.

This changeset allows TTL to be any of these:
 - None: turn off caching
 - 0: Caches forever (Default)
 - number: seconds before cache invalidation. Should be positive
 - tuple `(ttl_min, ttl_max)`: Each value will take between `ttl_min` and `ttl_max` seconds to expire (inclusive). Ensure `1 <= ttl_min <= ttl_max`
 - no-argument function: Will be called before every cache insertion to determine the TTL of the new entry. Should return a positive number