Support for dict-like data stores
Using `@memorized()` is a simple and effective way to add caching to an application.

But on some applications, we might not want to store our data in memcached.

A few simple examples: Using a plain `{}` cache is good solution if we don't need to share the cache between multiple executions, and `shelve`s are a good way to store results in disk.